[refactor] 역할별 폴더 구조 변경
