+++ src/pages/RestaurantPage.jsx
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 
-import RestaurantContainer from './RestaurantContainer';
+import RestaurantContainer from '../containers/RestaurantContainer';
 
 export default function RestaurantPage({ params }) {
   const { id } = params || useParams();
